subscription: Add the ParseAttachedSubscriptionsTask
The task parses attached subscriptions as well as final
system purpose data from the installation environment
after a successful subscription.

(System purpose data can change if an activation key
with attached system purpose data is used. Due to this we
need to parse system purpose data after subscription has
been attached to avoid the Subscription spoke showing stale
system purpose data after a subscription has been attached.)

We connect to the completed signal of the task to automatically
set attached subscriptions and system purpose data when the task
succeeds. And we do the same in reverse for the unregister task.