Change return of SetUpHardDriveSourceTask
Return namedtuple instead of two value tuple which is better to read.
The second tuple part is changed from bool to string and it will return
name of the iso image. We need this for HardDrive source module API.
It's replacing meaning of the boolean variable because the empty string
means the same as the bool.