subscription: Add asynchronous registration helper functions
Add two helper functions for asynchronous registration & unregistration.
The functions are meant to be called in a thread and provide feedback
as well as error reporting to the caller via callbacks.

During normal operations the functions call the progress callback with
values specified by the SubscriptionPhase enum. On success SubscriptionPhase.DONE
phase is set and the helper function terminates. On error, the error callback
is called with appropriate error message and the helper function terminates.

It is expected these helper functions will be started in a thread, as the work
they are doing (especially in the register & subscribe) case can take in extreme
cases tens of seconds or even up to a minute.

Also add comprehensive unit tests that make sure the rather complex
helper functions do what they should & handle expected error states.