subscription: Add IsRegistered property
This property is needed to detect if a system is currently registered
but does not have a subscription attached. This usually signifies an
error state, usually caused by the attempt to attach subscription
to the system failing for some reason.

The new property is also hooked to the DBus tasks:
- registration tasks set it to True on success
- unregistration task sets it to False on success