subscription: Make it possible to register/unregister from the GUI
Make is possible to register & subscribe or to unregister from the
GUI in the Subscription spoke. Unlike with kickstart, it is
possible to use not just org & key but also username & password
authentication.

The registration & subscription helper function is started in
a "subscription" thread, so that it does not block the GUI
when we wait for RHSM to complete the given operation.

Callbacks are used to provide feedback, report errors and
update spoke state.