subscription: Show data about attached subscriptions
This basically updates the second part of the spoke that
is shown after a successful subscription attempt. Type
of authentication used, final system purpose data as
well as information about attached subscriptions is shown.

The code that creates the listbox delegates describing an
attached subscription is not very nice, as it needs to
imperatively construct the rather complex listbox delegate
using the GTK/Pango API.

Ideally this would be done be describing the delegate in
Glade and then just instantiating it per item and filling
it with data. Unfortunately there is zero information
available for dynamically creating widgets with Glade
and I was simply not able to get it working in this way.

I guess this might explain why some of the dynamically
created widgets we have in Anaconda are implemented in C.

Also add definitions to the anaconda-gtk.css file to make
the attached subscriptions listbox look as intended.