Update the bootloader configuration after live installation
The GRUB configuration was changed to not store the kernel cmdline as a
variable in the grubenv file anymore, but instead in the BLS snippets.

But this change broke the live installations because the kernel-install
grub2 plugin now takes the kernel cmdline from /proc/cmdline. That means
the BLS snippets will have the cmdline used to boot the live install and
not the one needed to boot the installed rootfs.

To make sure that the BLS files have the expected cmdline, call to the
grub2-mkconfig tool after all the kernel have been installed.

Resolves: rhbz#1836020

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>