Migrate set NFS of SourceSwitchHandler to source modules
Make the method complete. Part of the NFS data was set outside of this
method and part inside. Make everything on one place.

Remove switching to NFS source before really setting NFS to be able to
set the data before setting NFS source. Don't know why this was there
really.