subscription: Make credential checking functions usable in GUI
The username_password_sufficient() and org_key_sufficient() helper
functions used to always fetch the subscription request on invocation
over DBus. This is not a problem for sporadic invocation, but would be
a much bigger problem for the Subscription spoke usecase, where this
function needs to be invoked basically for any character typed to the
authentication fields.

So lets make it possible to provide the checking functions with a
SubscriptionRequest instance that they can use instead of fetching
one over DBus. Also adjust tests for the helper functions to cover
this usecase, including the change to also consider SECRET_TYPE_TEXT
as being set, not just SECRET_TYPE_HIDDEN.