Convert help system to use help id
Instead of separate help files per Anaconda screen there is now a
single Docbook document and we tell Yelp to access individual sections
of the document via so called "anchors".

There is a separate JSON file that we parse that provides mapping
between individual screens (each with a unique help id) and anchors.
This is in a separate file to allow anchors in the documentation to
change without the need to make changes in Anaconda at the same time.

Also handle the difference between system language being specified as
locale code (for example en_US) and localized help content being stored
in directories named after the language code (en-US).

Related: rhbz#1691319