subscription: Initial CDN source switching
Implement initial support for switching to the CDN as
installation source, both for automatic kickstart installation
as well as for manual GUI based installation.

This is achieved by adding a couple helper functions
that make it easy to check if CDN is the current installation
source as well as setting CDN as installation source,
then calling these in the main register_and_subscribe()
helper function.

In this initial form we simply always set the CDN as installation
source after a successful registration+subscription. Checking
if it is suitable to override the existing installation source
as well as switching back to a more suitable installation source
after unregistration or on error is not yet implemented.

Also make sure to restart the payload if we expect the repo files to
change. We do not need to do that for automated installation, in
this case we make the payload restart thread wait for the subscription
thread instead.