Remove the changed property of the Source spoke in GUI
Use the default property that always returns True and move the code to the
method apply. The method apply will reset the payload if anything has changed.

Otherwise, the property could cause tearing down of the source without setting
it up again in the apply method. Then the installation started with unmounted
sources.