Do not mount harddrive sources as RO
It is correct way to do. To mount installation sources as RO only. We
don't want to modify them in any way and this will prevent us to do
that. However, we are also mounting the sources in SourceSpoke as RW.

Unfortunately, kernel does not allow to have one source mounted as RO
and RW at the same time.