subscription: Only require Insights package if subscribed
Red Hat Insights is enabled by default in the GUI, even if
the user never enters the Connect to Red Hat spoke.

This together with custom repo URL being used will result
in the Insights package being added to the transaction,
even though it can't really be used without the system having
a valid subscription attached.

So before adding the Insights package as a requirement, check
that the system is subscribed.