subscription: Make sure CDN repos are loaded and usable
We need to do two things to make the CDN provided repositories
available for installation:

1) we need to make sure the CDN provided repos loaded from the
   RHSM generated redhat.repo file

Normally no repo file are loaded for kickstart installations to
prevent the closest mirror repo files from interfering with
automated installations.

But if the CDN is set as the installation source and we are
subscribed, we want to override this and load the RHSM generated
repo file.

2) we need to make sure a base repo is available if CDN is used
   as the installation source

On Fedora for an installation from closest mirror we check
the available repos for a whitelist of well known repo names
and select one of them as base repo if it matches this whitelist.

In comparison, the repo names on the Red Hat CDN is no concept
of such stable well known repo names and the names are effectively
considered to be an implementation detail. Still, any enabled
repo in the redhat.repo file should be usable as a base repo.

So for the CDN, we just pick one of the enabled CDN repos to prevent
the payload code from falling over due to no base repo being set.

Based on commit eafef5ab85faa52b2efea3ad6526e12426551692.

Related: rhbz#1691319