Don't set up sources in the refresh method of the Source spoke
Use the current source to set up the GUI elements in the refresh method of
the Source spoke, but don't change it. The source should be changed only in
the apply method.

Set up the spoke for all supported types of sources and raise an exception
if the current type is unsupported.