Change the default source to CDROM
When we restart the payload thread for the first time, we should change
the default source to CDROM if there is a valid install media.

Anaconda uses the argument checkmount to control this behavior, but we
should try to find a better way how to handle this case.

Related: rhbz#1691319