subscription: Manual CDN selection support
Add a "Red Hat CDN" radio button to the source spoke,
so that users can explicitly select that CDN should
be used as the installation source. It also indicates
CDN is the current installation source and makes it
possible to seamlessly switch to a different source.

This requires not just adding the CDN radio button but
also making sure it is toggled appropriately based
on current installation source and tweaking the status
messages of the source and software spokes.

Also make sure the button is hidden if the Subscription
DBus module is not available.

Related: rhbz#1845605