subscription: Handle cases where CDN should not be the default
The CDN is the default source but if on media repo is found,
override the CDN and use the on media repo as the installation source.

The same thing for user specified repo URL, that also has priority over
the CDN.

Previously this override only worked for the DNF payload default source,
now we need to teach it to also consider the CDN as another possible
default source.

NOTE: After a successful registration the CDN will be automatically
      set as the installation source even if on media repo was found.
      The rationale for this is that the CDN will always have the
      same or (more likely) newer and more secure packages than
      the on media repo. Using the CDN if available also reduces the
      amount post installation package updates that have to be done
      on the system.

Related: rhbz#1845605