subscription: Handle source switching at registration/unregistration
So far we have simply set the CDN as installation source after
every successful registration.

This is simple but not optimal as it does not handle all valid
use cases.

So change the logic to only set the CDN as installation source if
the current source is on a list of appropriate sources to be overridden
by the CDN and remember what was the original source.

In a similar manner, roll back from CDN as installation source if
the original source was on the list of sources to be overridden.

At the moment the list of sources to be overridden only contain the
CDROM source, as we consider on media repos outdated by default.

Related: rhbz#1845605