Do not test if repo is valid based on .treeinfo file (#1844287)
Not all repositories need to have .treeinfo file. When it is not a compose but
only a third party repo it's probably created by just running createrepo_c which
does not create this file. We do not want to disable these repositories.

So instead check that repodata/repomd.xml file is present. Based on my
discussion with DNF/RPM developers it seems like the best approach.

Resolves: rhbz#1844287
Resolves: rhbz#1849093

Reported-by: Adam Williamson <awilliam@redhat.com>