subscription: Introduce the default_source configuration option
Add the default_source option to the Payload section of the
Anaconda configuration file. This option introduces initial support
for setting the default installation source via the configuration
file.

At the moment we only document the CLOSEST_MIRROR and CDN sources as
valid, due to the main aim if this change being easy switching of
default DNF payload source. In the future this can be extended to
cover also to the other sources as necessary.

Related: rhbz#1845605