Fix software spoke message when source changes
With the Red Hat CDN now easily available the users will now
trigger the software spoke message that tells them to re-enter
the software spoke and confirm the software selection.

Unfortunately the current message - "Nothing selected" is
rather confusing. It does not tell what is not selected and
might actually be incorrect, as in some cases the previous
selection might transfer over when the installation source
changes.

So change the confusing message to "Please confirm software
selection" which is also more in line with our other spoke
status messages.

Thanks a lot to Jan Stodola for coming up with the text of
the new message. :)

Resolves: rhbz#1788458
(cherry picked from commit 6867d6eef6ce777c2059fe3ae278c858790c5cf7)