Remove the encrypted attribute
The encrypted attribute of the default_partitioning configuration option
can be removed. It will be allowed to encrypt all mount points specified
in the Anaconda configuration files. It won't affect boot partitions that
are defined somewhere else. It will affect the default partitioning of Red
Hat Virtualization and oVirt Node Next, but we are able to with deal that.