Add workaround to fix support for installtree without repo
Sources should always point to the repository, however, right now we are in
middle of the migration process to Payloads sources so only base
repositories are migrated and they are more or less only reflecting user data.
To solve this for now we are adding back support for .treeinfo file and not
just repodata. This should be removed when payload source migration will be
finished.

Resolves: rhbz#1854825