Wait for payload initialization to finish in Subscription spoke
When the Subscription spoke first starts up wait for the
payload initialization thread to finish.

Otherwise, there is a chance incorrect initial spoke status will be set
for the Subscription spoke. This is due to the fact that the payload
initialization thread might find a CDROM source and set is as the
default installation source. If Subscription spoke sets its initial
state before that, it is possible for the state to be incorrect,
such as the spoke being mandatory even when CDROM is the installation
source.

Related: rhbz#1823578
(cherry picked from commit 91a2719951661adfdf81af91116a0094fde5d085)