Move slower part of Subscription spoke initialization to a thread
A part of the Subscription spoke initialization could be potentially
slower, as it waits for other initialization threads to finish.

So move that part of initialization to a separate thread, as is
done already in other spokes for the same reason. Also make the
spoke not ready, until the initialization thread (and thus also
the initial payload refresh) runs to the end.

As this effectively moves some GUI related method calls into
a thread, decorate the called method with the appropriate
async decorator, to avoid GTK weirdness due to calls from a thread.

Related: rhbz#1823578
(cherry picked from commit 11aaa66098c1603bb0679a13ec2cf901f18473fb)