Discard current boot list when updating the boot-device NRVAM variable
The boot-device NVRAM variable contains the list of devices that should be
used to boot a PowerPC-64 machine. On installation the device used for the
PreP partition is stored there.

But when doing that, the previous boot list is added after the device that
contains the PreP partition to retain the existing boot devices.

This could lead to the system failing to boot, even when the installation
finished successfully. For example if there were already wrong values set
in the boot-device variable, the firmware will fail to parse the variable
and fail to boot the machine.

One option is to use the SMS menu to restore the Boot Device Order to the
default values. But it's more robust if the installer could always work
even when the boot-device NVRAM variable had some wrong values set.

So instead of attempting to carry over the current boot list, just ignore
the current value and set the OF device paths for the installed partition.

This also makes sense because there is a limit on how many devices can be
stored in the boot-device variable, and on multipath setups it's possible
that the multiple device paths needed will take all the allowed entries.

Resolves: rhbz#1854307

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>