Use bootlist command to update the PowerPC-64 Boot Order
The boot-device NVRAM variable is used to store the Boot Order used by the
PowerPC-64 firmware to know what device should be used. The boot devices
are defined using the full Open Firmware device paths.

Currently the installer first calls the ofpathname command to translate a
logical Linux device used for the PreP partition to the OF device paths,
and then calls to the nvram tool to update the boot-device NVRAM variable.

But there's a bootlist tool that should be used for this instead. Since it
already takes care of doing the logical device to OF path translation and
checks that the number of devices doesn't exceed the maximum allowed.

Resolves: rhbz#1854307

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>