CDRom source should prioritize stage2 device during discover
We should first try to look on device used for the boot and then continue to
look on other devices. Return back this behavior. This is especially required
for booting from DVD ISO where inst.stage2 is used.

Also add TODO to return back support for cdrom source for a specific device
and use that in the inst.stage2 for the DVD ISO images to simplify everything.

Resolves: rhbz#1856264