Reload treeinfo repositories on every payload reset
Remove old repositories before loading new ones. We are changing the logic to
load new ones every time the base repo is changed.

This will solve problem that additional repositories pointing to an
invalid path. This is happening because we moved source mounting to a
different folders for each source.

Resolves: rhbz#1851207