Remove treeinfo repositories instead of disabling
We now remove repositories from treeinfo file on each payload reset. In that
case we should not disable these in the source spoke but instead remove them.

We should ideally also show them back to user when the source is switched back
but that logic to detect if base repo repository have changed is not trivial and
we had to change _update_payload_repos to achieve this. We don't want to
introduce any more regressions and that code change could do that.

Related: rhbz#1851207