Implement get_source_proxy() in payload base class
Even though some payloads (such as for example the OSTree payload)
do not have a source proxy, it's better to implement this method
universally. We can just return None for the payloads that don't
have a source.

As the method now can return None, adjust the code that can be using
it with a payload that does not have source proxy to handle this case.
We do this by checking if the payload is the DNF payload and if it is,
we assume it will return a valid source proxy.

Resolves: rhbz#1861456
(cherry picked from commit e3144e972703fac6573e6dd3cb1a46bbdce8728b)