subscription: Convert the RHSM default config values to expected format
Turns out the RHSM config DBus API GetAll() returns a nested
distionary instead of a flat dict we were expecting, as SetAll()
uses such a flat dict.

Due to this empty values from GUI would be written to the RHSM
config file as no original values would be found due to the
unexpected format of the GetAll() returned dict.

To fix this, convert the nested dict to the flat one we expect &
which is much easier to work with in our use case.

This way the Candlepin URL & CDN URL should no longer be overwritten
by empty values from the GUI.

Resolves: rhbz#1862116