Remove the DBus method ConfigureNTPServiceEnablementWithTask
Don't enable or disable the NTP service via the Services DBus module. Extend
the installation task ConfigureNTPTask of the Timezone module and configure
the NTP service there.

Change the order of the installation tasks of the Services and Timezone module,
so the NTP service can be always disabled with 'services --disabled=chronyd'.

It is not necessary to skip the requirement of the Timezone module if the
chrony package is excluded from the installation package set, because we do
that automatically for all collected requirements.

(cherry picked from commit 7db251aaeb826ef0c120fec87971dee330f2d3f6)