subscription: Only restart payload when needed
Only restart the payload after registration/unregistration
when necessary - eq. when CDN is in use or when CDN overrides
another installation source (CDROM).

Do not restart the payload in cases when it is not needed,
such as when the URL installation source is in use.

To do this, move the payload restart logic to the subscription
helper thread, and make it conditional, as payload restart is
currently not needed during registration triggered by an
automated kickstart installation, only during manual interaction
in the GUI.

Resolves: rhbz#1861351