Propagate verify_ssl to RHSM
At the moment the inst.noverifyssl boot option and the corresponding
Anaconda config file option verify_ssl only influence SSL certification
validation when anaconda itself downloads some installation related
artifacts.

For consistency this option should be also propagated to the RHSM
install time configuration.

Also this way users that acknowledge the risks by passing the
inst.noverifyssl boot option can successfully register against
a Candlepin instance with certificate that the installation
environment considers unknown or invalid.

Due to how this option is set (effectively overwriting the RHSM config
file) we need to make RHSM SSL certificate validation is turned back ON
at the end of the installation. This is again consistent with how
SSL certificate verification overrides are implemented elsewhere in
Anaconda. The overrides are always install time only and never propagate
to the installed system.

On the impelementation side we rename one of the installation tasks
as it no longer only reverts the log level changes but also makes
sure SSL certification validation is enabled in RHSM on the installed
system. Also change from Set() to SetAll() as se no longer just set a
single value in the two tasks that now interface with the RHSM SSL
certificate validation toggle.

Resolves: rhbz#1858126