Avoid unnecessarily pulling in glibc-langpack-en
Back in 2016[1], this was changed to a strict dependency because the
releng tools of the time could not handle boolean dependencies in
spec files. Nowadays, these work just fine, so we should switch
back to ensuring that we only pull it in when glibc-all-langpacks
is not installed.

Related to https://pagure.io/releng/issue/9689

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1323314

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>