Clean up s390 partitioning check (#1855724)
a) The single_pv property does not exist in blivet since ca. 2004. See
   blivet's commits e4549cf and b7df921.

b) The complicated restrictions on /boot are better solved by relying on the
   general check across all architectures.

c) The comments about root were all wrong.

Deleting both that check and comments related to it resolves the issue. Only
the LDL check remains now.

Resolves: rhbz#1855724

Co-authored-by: Jan Stodola <jstodola@redhat.com>