Fix dependency_solver failure with spec file boolean logic syntax
There is proposed change to have requires in spec file like

Requires: (... or ...)

Problem with this is that we take package name with '('. To avoid this problem,
remove the brackets before processing and take the first item. Taking the first
is not the best solution but we don't have all the info to decide so go with
first and hope is our way now. It really should be fine in most of the cases.