Do not mount as RW in Dracut
Dracut do not have reason to mount sources as RW it's more that it wasn't
specified and until recently we just did not bother. However, we changed the
logic that installation sources in stage2 environment does not re-using mounts
from the Dracut so this will now fail to set the environment because you can't
mount one thing as RW and RO at once.

Resolves: rhbz#1871049