Fix traceback when removing additional repository
I made a mistake by missing `not` in the condition. That way the repo removal as
user action crashed Anaconda because variable which was not set was immediately
used.

Unfortunately, I tested this just on the failing use-case and in that case it
looked like it is working correctly (but didn't really).

Resolves: rhbz#1871037
(cherry picked from commit a92428aa61f31d3f25786d4e90108d0d7751d680)