Add test to detect every RW mount command in Dracut
We changed a logic how to source modules so each module is mounting it's on
source to it's own created folder. If anything is mounted as RW in Dracut it
will fail to be mounted in sources. This works correctly if all mounts are RO.

This test will check that each mount command is specified with '-o ro' otherwise
it have to be ignored.