network: default to addr-gen-mode eui64
With NetworkManager in initramfs the ifcfg files for devices activated in
initramfs are now created after switch root by Anoconda via NM API which
defaults to stable-privacy in this case so we need to explicitly set eui64.
Eui64 used to be the default when ifcfg files were created by legacy network
module - there were no value for IPV6_ADDR_GEN_MODE set in ifcfg file which
means using eui64 for NM (for compatibility).

Resolves: rhbz#1873021