subscription: Assure payload restart on DVD install after registration
Make sure that after successful registration on an install from a DVD image
Anaconda does not just switch the installation source to the CDN (as it
has more up to date and secure packages than the on media repos), but
also restarts the payload. Otherwise we can end in an in-between state
with CDN source being assumed, yet most things still pointing to the
on media repo, resulting in an install-time crash.

To fix this, make sure to get a new source proxy - the previously used
one is no longer valid after a source switch.

Also fix unit tests accordingly.

Resolves: rhbz#1873164