Always clear treeinfo metadata (#1872056)
Metadata from the treeinfo were loaded only during the load of new metadata.
However, this does not work if we have source without metadata (e.g.
mirrorlist). In that case we are loading additional repositories from the old
metadata (not mounted anymore) and not the new ones which may have unexpected
results.

Resolves: rhbz#1872056