Fix the combo box for an URL type of additional repositories (#1879127)
In the commit ff9a7e1, we introduced new constants for the URL source types with
values 'BASEURL', 'MIRRORLIST' and 'METALINK'. In the commit cc2c3d3, we started
to use these constants in the Installation Source spoke and removed the old ones
with values 'url', 'mirrorlist' and 'metalink'. We updated the combo box for the
base repository, but forgot to update the combo box for additional repositories.

The combo box items have to have ids that match the values of the constants,
otherwise the URL type of additional repositories will be always 'BASEURL'.

Resolves: rhbz#1879127