fix remove unkown partition in sda failed
fix: cannot create partition when sda exists a ext4 filesystem.
When you clicked the "-" button, only the format on sda should be destroyed without removing sda from the device tree, but sda was also destroyed. As a result, sda cannot be found during disk initialization and an error was reported. 
Resolves: rhbz#1878661