Mock system operations in MountFilesystemsTask task
That test previously ran lsblk and mount in the test environment. This
does not work in containers [1][2], is meaningless as there are no
"real" block devices during the unit test, and it's also inappropriate
and dangerous for a unit test.

Mock out the calls, and just assert that setting up /mnt/sysimage
happens as expected in blivet and pyanaconda.core.

Now the test works as unprivileged user.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1884283
[2] https://github.com/containers/podman/issues/3112

Cherry-picked from master commit 645acef51.

Related: rhbz#1885635