Run rpm tests in a GitHub action
This replaces the "Anaconda Rawhide test RPM files" Jenkins job. (When
that gets removed, a lot of Makefile rules can get cleaned up.)

As with the unit tests, use a standard Fedora rawhide container for now,
and install the anaconda build dependencies into it. Like with
unit tests, this should eventually be replaced with a custom container
image that gets refreshed regularly (gated by tests), to increase both
speed and reliability. But the Jenkins job has the exact same problems,
so this is not a regression.