Separate RPM installability test from rpm_tests
tests/rpm_tests mostly inspects the built rpms, which neither requires
any privileges nor apply any changes to the system.

`InstallRPMTestCase` stands out, as it assumes that the installation
happens in mock (which is not true any more when testing in containers
or a local build environment). It also requires a lot of code/mechanics.

Move the installability test into the `make test-rpm` top-level rule
(which is the one called by Jenkins), and drop `InstallRPMTestCase`.