Add GitHub actions runner to ci-tasks container
This allows us to start a self-hosted runner pretty much anywhere,
including locally (for debugging) or on internal infrastructure (for
RHEL tests).

Include a `github-action-run-once` script that does everything that's
required for this, so that the container invocation becomes fairly
simple (see comment at the top of the file).

Move the default entrypoint to /, to keep all entry points at the same
destination directory, and make that as easy as possible to type.