Translate spoke category title directly at runtime
At the moment we mark spoke category title stored in class attribute for
translation and then translate it when hub sets up the category widget.

This had to be done as the class attribute is evaluated on code parsing
time and Anaconda might have a different locale (due to user selecting
installation language) by the tame the string is being displayed &
We would then translate the string at runtime.

This is unfortunately problematic for Initial Setup and addons as there
the category title will be marked for translation in the IS/addon
context and then the translation will be attempted in the Anaconda
context, which of course does not have the translated string.

So lets drop this split and instead translate the category title always
at runtime by providing it via a static class method. IS/addons can then
do the same, making sure the category title will be always translated at
the appropriate time & in their translation context.

This has also the side effect that the translated category titles which
are used for category sorting in TUI might result in the TUI category
order changing based on selected language. This would be quite
confusing, so instead of using alphabetical sort to decide category
order in the TUI just use the category sort order as is already done
in the GUI.

Also clean up and document the category widget handling code a bit.